pyproject.toml
[project]
name = "langchain-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "huggingface-hub>=1.1.7",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.22",
    "langchain-core>=0.3.80",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.35",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=0.2.5.1",
    "streamlit>=1.52.0",
    "transformers>=2.3.0",
]
